+++ CHANGELOG.md
@@ -2,3 +2,9 @@
 
 -  **Added:**
    -  New class EventEmitter
+
+## Version 0.0.5 (Upcoming)
+
+-  **Changed:**
+   -  Obj.omit: keep the object reference pointer
+   -  EventEmitter.emit: using handler.call(this, ...args)